estilos finalizados, o eso creo

+++ src/app/components/pay/pay.component.ts
@@ -2,8 +2,7 @@ import { Component, OnInit } from '@angular/core';
 import { Router } from '@angular/router';
 import { DetalleOrden } from '@app/shared/interfaces/detalleorden.interface';
 import { CarritoService } from '@app/shared/service/carrito.service';
-import { ToastrService } from 'ngx-toastr';
-
+import Swal from 'sweetalert2';
 @Component({
   selector: 'app-pay',
   templateUrl: './pay.component.html',
@@ -15,7 +14,6 @@ export class PayComponent implements OnInit {
 
   constructor(
     private carritoServ: CarritoService,
-    private toast: ToastrService,
     private router: Router
   ) { }
 
@@ -30,10 +28,28 @@ export class PayComponent implements OnInit {
   }
 
   payFood(): void {
-    this.toast.success('Hizo una compra ficticiosa :)', 'Felicidades');
-    this.total = this.carritoServ.deleteCart();
-
-    this.router.navigate(['']);
+    Swal.fire({
+      title: '<h2 class="sweetalert2-custom-title">¿Eso es todo?</h2>',
+      icon: 'info',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: '<h3>Si, es todo</h3>',
+      cancelButtonText: '<h3>Aun no</h3>',
+    }).then((result) => {
+      if (result.isConfirmed) {
+        Swal.fire({
+          title: "<h2 class='sweetalert2-custom-title'>Felicidades, acaba de hacer una compra ficticia :')</h2>",
+          icon: 'success',
+          confirmButtonText: '<h3>Cerrar</h3>',
+        }).then(() => {
+          if (result.isConfirmed) {
+            this.total = this.carritoServ.deleteCart();
+            this.router.navigate(['']);
+          }
+        });
+      }
+    });
   }
 
 }
